requirements-dev.txt
pytest==7.4.3
pytest-cov==4.1.0
bandit==1.7.5
safety==2.3.5
requests==2.31.0
flask-testing==0.8.1
